Splitted map series into several files.
