+++ ts/Series/Map/MapSeriesOptions.d.ts
@@ -1,0 +1,42 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ColorType from '../../Core/Color/ColorType';
+import type MapPointOptions from './MapPointOptions';
+import type MapSeries from './MapSeries';
+import type {
+    PointOptions,
+    PointShortOptions
+} from '../../Core/Series/PointOptions';
+import type ScatterSeriesOptions from '../Scatter/ScatterSeriesOptions';
+import type { SeriesStatesOptions } from '../../Core/Series/SeriesOptions';
+
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface MapSeriesOptions extends Highcharts.ColorSeriesOptions, ScatterSeriesOptions
+{
+    data?: Array<(PointOptions|PointShortOptions|MapPointOptions)>;
+    nullColor?: ColorType;
+    nullInteraction?: boolean;
+    states?: SeriesStatesOptions<MapSeries>;
+}
+
+export default MapSeriesOptions;
